separate SegmentTree into its own class and remove config min_frame_size_use_segment_tree
